requirements.txt
-i https://pypi.org/simple
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
github3.py==3.2.0
idna==3.3
pycparser==2.21
pyjwt[crypto]==2.4.0
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
